Revert "Merge pull request #786 from jherico/feature-vcpkg"
This reverts commit 198ed9d4b44c3fee2cb0a6071d35b4f91af625b2, reversing
changes made to 09c1eb76b4f3313ac8517b7c9dea70af6b5c8344.

+++ Cesium3DTilesContent/src/PntsToGltfConverter.cpp
@@ -5,7 +5,6 @@
 #include <CesiumGltf/ExtensionCesiumRTC.h>
 #include <CesiumGltf/ExtensionKhrMaterialsUnlit.h>
 #include <CesiumUtility/AttributeCompression.h>
-#include <CesiumUtility/Log.h>
 #include <CesiumUtility/Math.h>
 
 #ifdef _MSC_VER
@@ -23,6 +22,7 @@
 #endif
 
 #include <rapidjson/document.h>
+#include <spdlog/fmt/fmt.h>
 
 #include <limits>
 
@@ -815,7 +815,7 @@ bool validateDracoMetadataAttribute(
   }
 
   auto type = MetadataProperty::getTypeFromNumberOfComponents(
-      static_cast<int8_t>(pAttribute->num_components()));
+      pAttribute->num_components());
   return type && type.value() == semantic.type;
 }
